Don't pass register to expression compiler
Previously, the expression compiler took either a target register or a
flag telling it to return a new StaticReference.  The goal was to
dereference references where they were created and this necessary
before I had a compile-time reference representation, but this made
things more awkward overall and made it particularly difficult to
implement call expressions, which need a reference from the call
expression if possible, but accept a value if not.

Now, the expression compiler simply returns either a StaticReference
or a target code expression and the caller is responsible for emitting
GetValue code.  This matches very closely with the evaluation
algorithms given in the specification.  Since emitGetValue can
statically distinguish references from values, the generated code is
essentially as efficient and compact as before.