Change calling convention to support constructor calls
This changes the calling convention yet again so that environment
constructor functions don't have to return anything to indicate their
support of the JSJS calling convention to the caller.  This was going
to cause problems if the function was invoked as a constructor
because, if a constructor returns an object, that object is used as
the value of the "new" expression.

Instead, the environment constructor just pushes itself on the world's
call stack and the caller always returns to the execution loop.  If it
was actually a native call, the execution loop will jump right back to
the caller; if it was a JSJS call, it will enter the called function.